+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/internal/TestIntegrationDefaultsPlugin.java
@@ -24,9 +24,11 @@ import com.liferay.gradle.plugins.test.integration.TestIntegrationTomcatExtensio
 import com.liferay.gradle.plugins.test.integration.tasks.SetUpTestableTomcatTask;
 import com.liferay.gradle.plugins.test.integration.tasks.StartTestableTomcatTask;
 import com.liferay.gradle.plugins.test.integration.tasks.StopAppServerTask;
+import com.liferay.gradle.util.Validator;
 
 import java.io.File;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.Callable;
 
@@ -87,6 +89,11 @@ public class TestIntegrationDefaultsPlugin
 				project,
 				TestIntegrationPlugin.SET_UP_TESTABLE_TOMCAT_TASK_NAME);
 
+		setUpTestableTomcatTask.setJaCoCoAgentConfiguration(
+			GradleUtil.getProperty(
+				project, "jacoco.agent.configuration", (String)null));
+		setUpTestableTomcatTask.setJaCoCoAgentFile(
+			GradleUtil.getProperty(project, "jacoco.agent.jar", (String)null));
 		setUpTestableTomcatTask.setAspectJAgent(
 			GradleUtil.getProperty(project, "aspectj.agent", (String)null));
 		setUpTestableTomcatTask.setAspectJConfiguration(
@@ -134,6 +141,13 @@ public class TestIntegrationDefaultsPlugin
 
 				@Override
 				public List<String> call() throws Exception {
+					String argLine = System.getProperty(
+						"app.server.start.executable.arg.line");
+
+					if (Validator.isNotNull(argLine)) {
+						return Arrays.asList(argLine.split(" "));
+					}
+
 					return tomcatAppServer.getStartExecutableArgs();
 				}