+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/elements/DefinitionElement.java
@@ -116,6 +116,8 @@ public class DefinitionElement extends PoshiElement {
 			line = line.trim();
 
 			if (line.length() == 0) {
+				sb.append("\n");
+
 				continue;
 			}