+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/elements/VarElement.java
@@ -18,7 +18,9 @@ import com.liferay.poshi.runner.util.Dom4JUtil;
 
 import java.io.IOException;
 
+import org.dom4j.CDATA;
 import org.dom4j.Element;
+import org.dom4j.Node;
 
 /**
  * @author Kenji Heigel
@@ -27,8 +29,6 @@ public class VarElement extends PoshiElement {
 
 	public VarElement(Element element) {
 		this("var", element);
-
-		initValueAttributeName(element);
 	}
 
 	public VarElement(String readableSyntax) {
@@ -46,6 +46,14 @@ public class VarElement extends PoshiElement {
 	}
 
 	public String getVarValue() {
+		if (valueAttributeName == null) {
+			for (Node node : Dom4JUtil.toNodeList(content())) {
+				if (node instanceof CDATA) {
+					return node.getText();
+				}
+			}
+		}
+
 		return attributeValue(valueAttributeName);
 	}
 
@@ -61,6 +69,12 @@ public class VarElement extends PoshiElement {
 			value = value.replace("Util.", "Util#");
 
 			addAttribute("method", value);
+
+			return;
+		}
+
+		if (value.contains("\n")) {
+			addCDATA(value);
 
 			return;
 		}
@@ -120,6 +134,14 @@ public class VarElement extends PoshiElement {
 			valueAttributeName = "value";
 
 			return;
+		}
+
+		for (Node node : Dom4JUtil.toNodeList(element.content())) {
+			if (node instanceof CDATA) {
+				add((CDATA)node.clone());
+
+				return;
+			}
 		}
 
 		try {