+++ modules/sdk/gradle-plugins-source-formatter/CHANGELOG.markdown
@@ -814,12 +814,20 @@ another task with the same name is being executed in a parent project.
 ## 2.0.38 - 2017-07-19
 
 ### Changed
-- [LPS-72705]: Update the [Liferay Source Formatter] dependency to version 1.0.427.
+- [LPS-72705]: Update the [Liferay Source Formatter] dependency to version
+1.0.427.
 
 ## 2.0.39 - 2017-07-20
 
 ### Changed
-- [LPS-73600]: Update the [Liferay Source Formatter] dependency to version 1.0.428.
+- [LPS-73600]: Update the [Liferay Source Formatter] dependency to version
+1.0.428.
+
+## 2.0.40 - 2017-07-21
+
+### Changed
+- [LPS-72705]: Update the [Liferay Source Formatter] dependency to version
+1.0.429.
 
 [Liferay Source Formatter]: https://github.com/liferay/liferay-portal/tree/master/modules/util/source-formatter
 [LPS-52675]: https://issues.liferay.com/browse/LPS-52675